moved the icons to the icons folder
