+++ src/index.scss
@@ -80,12 +80,6 @@ button {
   text-transform: uppercase;
 }
 
-//Input
-input {
-  font-size: 12px;
-  font-weight: 700;
-}
-
 //Caption
 span {
   &.big {
@@ -136,3 +130,19 @@ main {
   position: relative;
   flex: 1;
 }
+
+//body::-webkit-scrollbar {
+//  background-color: #fff;
+//  width: 16px;
+//}
+//body::-webkit-scrollbar-track {
+//  background-color: #fff;
+//}
+//body::-webkit-scrollbar-thumb {
+//  background-color: #babac0;
+//  border-radius: 16px;
+//  border: 4px solid #fff;
+//}
+//body::-webkit-scrollbar-button {
+//  display:none;
+//}
