+++ src/pages/Artist/index.ts
@@ -1,0 +1,1 @@
+export { default as Artist } from "./Artist";
