+++ src/stories/IconButton.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { IconButton } from "@ui/IconButton";
-import { ReactComponent as Delete } from "@assets/delete.svg";
-import { ReactComponent as Moon } from "@assets/moon.svg";
-import { ReactComponent as Sun } from "@assets/sun.svg";
-import { ReactComponent as ArrowUp } from "@assets/arrow-up.svg";
+import { ReactComponent as Delete } from "@assets/icons/delete.svg";
+import { ReactComponent as Moon } from "@assets/icons/moon.svg";
+import { ReactComponent as Sun } from "@assets/icons/sun.svg";
+import { ReactComponent as ArrowUp } from "@assets/icons/arrow-up.svg";
 
 const meta: Meta<typeof IconButton> = {
   component: IconButton,
