+++ src/ui/Card/Card.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
 
 import { Link, LinkProps } from "react-router-dom";
-import { ReactComponent as ArrowNext } from "@assets/arrow-next.svg";
+import { ReactComponent as ArrowNext } from "@assets/icons/arrow-next.svg";
 import classNames from "classnames/bind";
 
 import styles from "./styles.module.scss";
