+++ src/ui/Input/index.ts
@@ -1,0 +1,1 @@
+export { default as Input } from "./Input";
