+++ src/ui/Overlay/style.module.scss
@@ -1,0 +1,17 @@
+.overlay {
+  transition: all 0.3s ease-out;
+  position: absolute;
+  z-index: -1;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(0, 0, 0, 0.3);
+  width: 100%;
+  height: 100%;
+  opacity: 0;
+  &-visible {
+    opacity: 1;
+    z-index: 100;
+  }
+}
