--- src/ui/Scrim/style.module.scss
@@ -1,17 +1,0 @@
-.scrim {
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  background-color: rgba(0, 0, 0, 0.3);
-  z-index: -1;
-  visibility: hidden;
-  opacity: 0;
-  transition: all 0.3s ease-out;
-  &.open {
-    opacity: 1;
-    z-index: 200;
-    visibility: visible;
-  }
-}
